requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
azure-identity>=1.17.0
azure-ai-projects>=1.0.0b11
pydantic>=2.7.0
